+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Feature: Added specs for the SimonGame
   * Bugfix: Generate sane spec defaults for new apps and controllers
   * Bugfix: `CallController#record` now allows partial-second timeouts
+  * Feature: Added the ability to specify a renderer to #play_audio and #play_audio!
 
 # [2.1.3](https://github.com/adhearsion/adhearsion/compare/v2.1.2...v2.1.3) - [2012-10-11](https://rubygems.org/gems/adhearsion/versions/2.1.3)
   * Bugfix: Originating call is now answered before joining calls using `CallController#dial`